- added the panel that would perform all the actions - added the event methods to navigate through the list (still buggy) - slider is now working with auto arrange - slider value gets updated - icons on the button are not being displayed!!!
