+++ input.txt
@@ -1,2 +1,5 @@
 600 300
-0
+3
+0.0,0.0 101.5992,0.0 101.5992,101.5992 0.0,101.5992
+0.0,0.0 102.81339,0.0 102.81339,135.20848 0.0,135.20848
+0.0,0.0 287.87833,0.0 287.87833,290.04 0.0,290.04
