--- src/com/t_oster/visicut/gui/ArrangeFrame.java
@@ -1,287 +1,0 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
-
-/*
- * ArrangeFrame.java
- *
- * Created on Feb 1, 2015, 5:01:31 PM
- */
-package com.t_oster.visicut.gui;
-
-import com.t_oster.visicut.gui.beans.PreviewPanel;
-import java.awt.geom.NoninvertibleTransformException;
-import org.jdesktop.application.Action;
-
-/**
- *
- * @author Marcel
- */
-public class ArrangeFrame extends javax.swing.JFrame
-{
-  int count;
-
-  /** Creates new form ArrangeFrame */
-  public ArrangeFrame()
-  {
-    initComponents();
-    count = 1;
-  }
-
-  /** This method is called from within the constructor to
-   * initialize the form.
-   * WARNING: Do NOT modify this code. The content of this method is
-   * always regenerated by the Form Editor.
-   */
-  @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-        bindingGroup = new org.jdesktop.beansbinding.BindingGroup();
-
-        numberOfArrangements = new javax.swing.JLabel();
-        stepContainer = new javax.swing.JPanel();
-        previousButton = new javax.swing.JButton();
-        nextButton = new javax.swing.JButton();
-        stepCount = new javax.swing.JLabel();
-        discardConfirmContainer = new javax.swing.JPanel();
-        discardButton = new javax.swing.JButton();
-        confirmButton = new javax.swing.JButton();
-        objectDistanceContainer = new javax.swing.JPanel();
-        distanceSlider = new javax.swing.JSlider();
-        distanceValueLabel = new javax.swing.JLabel();
-        rearrangeButton = new javax.swing.JButton();
-        seperator = new javax.swing.JSeparator();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setName("Form"); // NOI18N
-
-        numberOfArrangements.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        org.jdesktop.application.ResourceMap resourceMap = org.jdesktop.application.Application.getInstance(com.t_oster.visicut.gui.VisicutApp.class).getContext().getResourceMap(ArrangeFrame.class);
-        numberOfArrangements.setText(resourceMap.getString("numberOfArrangements.text")); // NOI18N
-        numberOfArrangements.setName("numberOfArrangements"); // NOI18N
-
-        stepContainer.setBorder(javax.swing.BorderFactory.createEtchedBorder());
-        stepContainer.setName("stepContainer"); // NOI18N
-
-        javax.swing.ActionMap actionMap = org.jdesktop.application.Application.getInstance(com.t_oster.visicut.gui.VisicutApp.class).getContext().getActionMap(ArrangeFrame.class, this);
-        previousButton.setAction(actionMap.get("previousArrangement")); // NOI18N
-        previousButton.setText(resourceMap.getString("previousButton.text")); // NOI18N
-        previousButton.setFocusable(false);
-        previousButton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        previousButton.setName("previousButton"); // NOI18N
-        previousButton.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
-
-        nextButton.setAction(actionMap.get("nextArrangement")); // NOI18N
-        nextButton.setText(resourceMap.getString("nextButton.text")); // NOI18N
-        nextButton.setFocusable(false);
-        nextButton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        nextButton.setName("nextButton"); // NOI18N
-        nextButton.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
-
-        stepCount.setText(resourceMap.getString("stepCount.text")); // NOI18N
-        stepCount.setName("stepCount"); // NOI18N
-
-        org.jdesktop.beansbinding.Binding binding = org.jdesktop.beansbinding.Bindings.createAutoBinding(org.jdesktop.beansbinding.AutoBinding.UpdateStrategy.READ_WRITE, this, org.jdesktop.beansbinding.ObjectProperty.create(), stepCount, org.jdesktop.beansbinding.BeanProperty.create("horizontalAlignment"));
-        bindingGroup.addBinding(binding);
-
-        org.jdesktop.layout.GroupLayout stepContainerLayout = new org.jdesktop.layout.GroupLayout(stepContainer);
-        stepContainer.setLayout(stepContainerLayout);
-        stepContainerLayout.setHorizontalGroup(
-            stepContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(stepContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .add(previousButton, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 58, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
-                .add(stepCount)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
-                .add(nextButton, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 56, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        stepContainerLayout.setVerticalGroup(
-            stepContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(stepContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .add(stepContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.CENTER)
-                    .add(previousButton)
-                    .add(stepCount)
-                    .add(nextButton))
-                .addContainerGap())
-        );
-
-        discardConfirmContainer.setName("discardConfirmContainer"); // NOI18N
-
-        discardButton.setAction(actionMap.get("discard")); // NOI18N
-        discardButton.setName("discardButton"); // NOI18N
-
-        confirmButton.setAction(actionMap.get("confirm")); // NOI18N
-        confirmButton.setName("confirmButton"); // NOI18N
-
-        org.jdesktop.layout.GroupLayout discardConfirmContainerLayout = new org.jdesktop.layout.GroupLayout(discardConfirmContainer);
-        discardConfirmContainer.setLayout(discardConfirmContainerLayout);
-        discardConfirmContainerLayout.setHorizontalGroup(
-            discardConfirmContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(discardConfirmContainerLayout.createSequentialGroup()
-                .add(25, 25, 25)
-                .add(discardButton)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED, 78, Short.MAX_VALUE)
-                .add(confirmButton)
-                .add(17, 17, 17))
-        );
-        discardConfirmContainerLayout.setVerticalGroup(
-            discardConfirmContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(discardConfirmContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .add(discardConfirmContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-                    .add(confirmButton, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, 33, Short.MAX_VALUE)
-                    .add(discardButton, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addContainerGap())
-        );
-
-        objectDistanceContainer.setBorder(javax.swing.BorderFactory.createTitledBorder(resourceMap.getString("objectDistanceContainer.border.title"))); // NOI18N
-        objectDistanceContainer.setName("objectDistanceContainer"); // NOI18N
-
-        distanceSlider.setMaximum(10);
-        distanceSlider.setValue(0);
-        distanceSlider.setName("distanceSlider"); // NOI18N
-
-        distanceValueLabel.setText(resourceMap.getString("distanceValueLabel.text")); // NOI18N
-        distanceValueLabel.setName("distanceValueLabel"); // NOI18N
-
-        rearrangeButton.setAction(actionMap.get("rearrangeObjectsBasedOnDistanceValue")); // NOI18N
-        rearrangeButton.setText(resourceMap.getString("rearrangeButton.text")); // NOI18N
-        rearrangeButton.setName("rearrangeButton"); // NOI18N
-
-        org.jdesktop.layout.GroupLayout objectDistanceContainerLayout = new org.jdesktop.layout.GroupLayout(objectDistanceContainer);
-        objectDistanceContainer.setLayout(objectDistanceContainerLayout);
-        objectDistanceContainerLayout.setHorizontalGroup(
-            objectDistanceContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(objectDistanceContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .add(distanceSlider, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 120, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
-                .add(distanceValueLabel)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
-                .add(rearrangeButton, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 87, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-        objectDistanceContainerLayout.setVerticalGroup(
-            objectDistanceContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(objectDistanceContainerLayout.createSequentialGroup()
-                .addContainerGap()
-                .add(objectDistanceContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.TRAILING)
-                    .add(objectDistanceContainerLayout.createParallelGroup(org.jdesktop.layout.GroupLayout.BASELINE)
-                        .add(distanceValueLabel)
-                        .add(rearrangeButton))
-                    .add(distanceSlider, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap())
-        );
-
-        seperator.setName("seperator"); // NOI18N
-
-        org.jdesktop.layout.GroupLayout layout = new org.jdesktop.layout.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(org.jdesktop.layout.GroupLayout.TRAILING, layout.createSequentialGroup()
-                .addContainerGap(56, Short.MAX_VALUE)
-                .add(layout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-                    .add(layout.createParallelGroup(org.jdesktop.layout.GroupLayout.TRAILING, false)
-                        .add(org.jdesktop.layout.GroupLayout.LEADING, discardConfirmContainer, 0, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .add(org.jdesktop.layout.GroupLayout.LEADING, stepContainer, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                    .add(layout.createSequentialGroup()
-                        .add(24, 24, 24)
-                        .add(numberOfArrangements)))
-                .add(46, 46, 46))
-            .add(seperator, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, 295, Short.MAX_VALUE)
-            .add(layout.createSequentialGroup()
-                .addContainerGap()
-                .add(objectDistanceContainer, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 255, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(org.jdesktop.layout.GroupLayout.LEADING)
-            .add(layout.createSequentialGroup()
-                .add(28, 28, 28)
-                .add(objectDistanceContainer, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .add(26, 26, 26)
-                .add(seperator, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, 10, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .add(numberOfArrangements)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.UNRELATED)
-                .add(stepContainer, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(org.jdesktop.layout.LayoutStyle.RELATED)
-                .add(discardConfirmContainer, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE, org.jdesktop.layout.GroupLayout.DEFAULT_SIZE, org.jdesktop.layout.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap())
-        );
-
-        bindingGroup.bind();
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-  /**
-   * @param args the command line arguments
-   */
-//  public static void main(String args[])
-//  {
-//    java.awt.EventQueue.invokeLater(new Runnable()
-//    {
-//
-//      public void run()
-//      {
-//        new ArrangeFrame().setVisible(true);
-//      }
-//    });
-//  }
-
-  @Action
-  public void discard()
-  {
-   
-  }
-
-  @Action
-  public void confirm()
-  {
-  }
-
-  @Action
-  public void previousArrangement() throws NoninvertibleTransformException
-  {
-//    --count;
-//    MainView.previewStaticPanel.navigateThroughArrangements(count);
-//    if (MainView.previewStaticPanel.)
-  }
-
-  @Action
-  public void nextArrangement() throws NoninvertibleTransformException
-  {
-    if(MainView.previewStaticPanel.isGreaterThanAllValues(count)){
-      discardButton.setEnabled(false);
-    }
-    MainView.previewStaticPanel.navigateThroughArrangements(count);
-    count++;
-  }
-
-  @Action
-  public void rearrangeObjectsBasedOnDistanceValue()
-  {
-  }
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton confirmButton;
-    private javax.swing.JButton discardButton;
-    private javax.swing.JPanel discardConfirmContainer;
-    private javax.swing.JSlider distanceSlider;
-    private javax.swing.JLabel distanceValueLabel;
-    private javax.swing.JButton nextButton;
-    private javax.swing.JLabel numberOfArrangements;
-    private javax.swing.JPanel objectDistanceContainer;
-    private javax.swing.JButton previousButton;
-    private javax.swing.JButton rearrangeButton;
-    private javax.swing.JSeparator seperator;
-    private javax.swing.JPanel stepContainer;
-    private javax.swing.JLabel stepCount;
-    private org.jdesktop.beansbinding.BindingGroup bindingGroup;
-    // End of variables declaration//GEN-END:variables
-}
