+++ src/com/t_oster/visicut/gui/MainView.form
@@ -969,6 +969,9 @@
         <Property name="text" type="java.lang.String" resourceKey="arrangeButton.text"/>
         <Property name="name" type="java.lang.String" value="arrangeButton" noResource="true"/>
       </Properties>
+      <Events>
+        <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="arrangeButtonActionPerformed"/>
+      </Events>
       <AuxValues>
         <AuxValue name="JavaCodeGenerator_InitCodePost" type="java.lang.String" value="arrangeButton.setEnabled(ifFilesLoaded());"/>
       </AuxValues>
