+++ src/com/t_oster/visicut/gui/beans/CreateNewThicknessDialog.java
@@ -152,7 +152,7 @@ public class CreateNewThicknessDialog extends javax.swing.JDialog
                             .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
                                 .addComponent(jLabel3)
                                 .addComponent(jLabel2, javax.swing.GroupLayout.Alignment.TRAILING)))))
-                .addContainerGap(18, Short.MAX_VALUE))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
