+++ src/com/t_oster/visicut/gui/resources/AutoArrangePanel.properties
@@ -1,0 +1,10 @@
+# To change this template, choose Tools | Templates
+# and open the template in the editor.
+
+objectDistanceContainer.border.title=Inter Object Distance
+rearrangeButton.text=Rearrange
+distanceValueLabel.text=0
+numberOfArrangements.text=0 Arrangements found
+previousButton.text=Prev
+stepCount.text=0/0
+nextButton.text=Next
